python/requirements.txt
# 核心依赖
pandas>=1.5.0
requests>=2.28.0
psutil>=5.9.0
openpyxl>=3.0.10
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyyaml>=6.0
mysql-connector-python>=8.0.30
psycopg2-binary>=2.9.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
schedule>=1.2.0
black>=22.0.0
flake8>=5.0.0
# 新增依赖
pycryptodome>=3.19.0  # AES加密
python-dotenv>=1.0.0  # 可选，环境变量
# 网络工具依赖
requests[socks]>=2.28.0  # 代理检测
# DNS查询
dnspython>=2.4.2 